Shared superclass for RS and oRS to remove duplicated functions. Correct RS and oRS to handle errors of empty documents.
