+++ setup.py
@@ -27,8 +27,6 @@ setup(
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
         'Programming Language :: Python :: 3',
-        #'Programming Language :: Python :: 3.7',
-        #'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
     ],
     description="OCTIS: a library for Optimizing and Comparing Topic Models.",
